Give mitsuku memory using cache
Note that the cache might not work because it is stored in a dictionary, which will disappear after the script ends.  An easy fix would be to save the cache to a file. Reference this for information on how to do so: http://cachecontrol.readthedocs.io/en/latest/storage.html